+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - [X] PlatformHelper: isModLoaded
 - [X] MapDataWrapper optional lazy loading
 - [X] server side configs reload on /reload command 
-- [ ] data packs
+- [X] data/resource packs
 - [ ] config array / map comments on fields brought to top if list empty
 - [ ] nbt data for tile, entity, item
 - [ ] quilt example mod